+++ admin/admin_template.php
@@ -64,6 +64,7 @@ $logo_path    = htmlspecialchars($cmsSettings['logo_path']    ?? '');
 				<div class="sidebar-submenu">
 					<a href="user_list.php">👥 Користувачі</a>
 					<a href="invite.php">✉️ Запрошення</a>
+
 					<a href="/admin/logs.php">📄 Переглянути логі</a>
 					<a href="/admin/SQLAdmin/phpadmin.php">🗄️ База даних</a>					
 					<a href="/admin/file_manager.php">📁 Файли</a>
@@ -94,6 +95,7 @@ $logo_path    = htmlspecialchars($cmsSettings['logo_path']    ?? '');
 				<div class="sidebar-submenu">
 					<a href="user_list.php">👥 Користувачі</a>
 					<a href="invite.php">✉️ Запрошення</a>
+
 					<a href="/admin/logs.php">📄 Переглянути логі</a>
 					<a href="/admin/SQLAdmin/phpadmin.php" target="_blank">🗄️ База даних</a>
 					<a href="/admin/file_manager.php">📁 Файли</a>